+++ CHANGELOG.md
@@ -4,7 +4,6 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog],
 and this project adheres to [Semantic Versioning].
-
 ## [0.1.0] - 2023-04-22
 
 - Initial release
@@ -85,3 +84,56 @@ and this project adheres to [Semantic Versioning].
 
 - Connection to the device is postponed now. Previously some out of range device might prevents HA from fully booting.
 - Improved connection stability.
+
+
+## [0.2.0] - 2024-03-21
+
+### Added
+
+- Add sfkzq/nxquc5lb device
+
+### Changed
+
+- Update readme (forked from)
+
+### Fixed
+
+- fix: Compatibility with HA 2024.1
+- Fix deprecated
+
+## [0.2.1] - 2025-03-26
+
+### Added
+
+- Add ggq/hfgdqhho device
+
+### Fixed
+
+- fix: Compatibility with HA 2025.3
+- Fix deprecated
+
+## [0.2.2] - 2025-04-16
+
+### Fixed
+
+- Fix deprecated with Home assistant 2025.4.2
+
+### Update
+
+- Update README about device support
+
+## [0.2.3] - 2025-04-25
+
+### Added
+
+- Add ggq/fnlw6npo device
+- Add jtmspro/ebd5e0uauqx0vfsp device
+
+## [0.2.4] - 2025-12-29
+
+### Added
+
+- drop CONF_APP_TYPE
+- Add mknd4lci
+- Add riecov42
+- fix SyntaxWarning: invalid escape sequence '\d'